log: avoid calling the logging functions for log levels above the current maximum
Messages with log levels above the current maximum log level will be dropped
inside log_meta(). But to be able to call the function, all parameters for
the function need to be provided. This can easily get expensive, if values
need to be calculated or functions are used in parameters.

Avoid all useless work by checking inside the macro, before we look
at any parameters passed to the logging functions.