logind: create the session fifo before saving the session file
If the session fifo is not created the session state written to
the session file is "closing". This caused the lock screen in
gnome-shell to go into a loop trying to find the active session.

The problem was introduced in the sd-bus port in
cc3773810855956bad92337cee8fa193584ab62e

Fixes: https://bugs.freedesktop.org/show_bug.cgi?id=71525